peak-revenues-advertising/app/requirements.txt
pip-tools
psycopg2-binary==2.8.4
environ-config
pylint==2.4.4
pandas==0.25.3
requests==2.22.0
unidecode==1.1.1
gspread==3.1.0
oauth2client==4.1.3
tabulate==0.8.6
conf==0.4.1